# docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Aura Render API",
    "version": "1.0.0",
    "description": "\n# Aura Render API Documentation\n\nAura Render is an AI-powered video generation platform that transforms text into professional videos.\n\n## Features\n- **AI Content Analysis**: Intelligent analysis of input text for optimal video generation\n- **Material Matching**: Automatic matching and acquisition of relevant video materials\n- **Real-time Processing**: WebSocket-based real-time progress tracking\n- **Multi-format Support**: Support for various video formats and quality levels\n- **Enterprise Security**: JWT authentication, rate limiting, and audit logging\n- **Batch Processing**: Handle multiple video generation requests efficiently\n\n## Authentication\nMost endpoints require authentication using JWT tokens. Include the token in the Authorization header:\n```\nAuthorization: Bearer <your-jwt-token>\n```\n\n## Rate Limiting\nAPI requests are subject to rate limiting:\n- **Free Users**: 10 requests/minute, 600 requests/hour\n- **Premium Users**: 100 requests/minute, 6000 requests/hour\n- **Enterprise**: Custom limits available\n\n## Error Handling\nAll errors follow a consistent format:\n```json\n{\n  \"error\": {\n    \"code\": \"ERROR_CODE\",\n    \"message\": \"Human-readable error message\",\n    \"details\": {\n      \"field\": \"Additional error context\"\n    }\n  }\n}\n```\n        ",
    "termsOfService": "https://aurarender.com/terms",
...
    },
...
    }
...
  "servers": [
...
  "security": [
    {
      "BearerAuth": []
    }
  ],
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT token obtained from /auth/login endpoint"
      }
    },
    "schemas": {
      "UserLogin": {
        "type": "object",
        "required": [
          "email",
          "password"
        ],
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "example": "user@example.com"
          },
          "password": {
            "type": "string",
            "format": "password",
            "minLength": 8,
            "example": "securePassword123"
          },
          "remember_me": {
            "type": "boolean",
            "default": false
          }
        }
      },
      "UserRegistration": {
        "type": "object",
        "required": [
          "email",
          "password",
          "first_name",
          "last_name"
        ],
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "example": "newuser@example.com"
          },
          "password": {
            "type": "string",
            "format": "password",
            "minLength": 8,
            "example": "securePassword123"
          },
          "first_name": {
            "type": "string",
            "maxLength": 100,
            "example": "John"
          },
          "last_name": {
            "type": "string",
            "maxLength": 100,
            "example": "Doe"
          },
          "company": {
            "type": "string",
            "maxLength": 200,
            "example": "Acme Corp"
          },
          "phone": {
            "type": "string",
            "example": "+1-555-123-4567"
          }
        }
      },
      "TokenResponse": {
        "type": "object",
        "properties": {
          "access_token": {
            "type": "string",
            "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
          },
          "refresh_token": {
            "type": "string",
            "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
          },
          "token_type": {
            "type": "string",
            "example": "Bearer"
          },
          "expires_in": {
            "type": "integer",
            "example": 3600
          }
        }
      },
      "UserProfile": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "usr_123456789"
          },
          "email": {
            "type": "string",
            "format": "email",
            "example": "user@example.com"
          },
          "first_name": {
            "type": "string",
            "example": "John"
          },
          "last_name": {
            "type": "string",
            "example": "Doe"
          },
          "role": {
            "type": "string",
            "enum": [
              "admin",
              "premium",
              "standard",
              "trial"
            ],
            "example": "premium"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "example": "2023-12-01T10:00:00Z"
          },
          "subscription": {
            "$ref": "#/components/schemas/Subscription"
          }
        }
      },
...
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "object",
            "properties": {
              "code": {
                "type": "string",
                "example": "VALIDATION_ERROR"
              },
              "message": {
                "type": "string",
                "example": "Invalid request parameters"
              },
              "details": {
                "type": "object",
                "example": {
                  "text": "Text content is required"
                }
              }
            }
          }
        }
      },
      "RateLimitError": {
        "type": "object",
        "properties": {
          "error": {
            "type": "object",
            "properties": {
              "code": {
                "type": "string",
                "example": "RATE_LIMIT_EXCEEDED"
              },
              "message": {
                "type": "string",
                "example": "Rate limit exceeded"
              },
              "details": {
                "type": "object",
                "properties": {
                  "limit": {
                    "type": "integer",
                    "example": 100
                  },
                  "remaining": {
                    "type": "integer",
                    "example": 0
                  },
                  "reset_time": {
                    "type": "integer",
                    "example": 1701432000
                  }
                }
              }
            }
          }
        }
      }
    }
  },
  "paths": {
    "/auth/register": {
      "post": {
        "tags": [
          "Authentication"
        ],
        "summary": "Register a new user",
        "description": "Create a new user account with email verification",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserRegistration"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User registered successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserProfile"
                }
              }
            }
          },
          "400": {
            "description": "Invalid registration data",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "Email already registered",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/auth/login": {
      "post": {
        "tags": [
          "Authentication"
        ],
        "summary": "User login",
        "description": "Authenticate user and get JWT tokens",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserLogin"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login successful",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenResponse"
                }
              }
            }
          },
          "401": {
            "description": "Invalid credentials",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "Too many login attempts",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RateLimitError"
                }
              }
            }
          }
        }
      }
    },
...
    "/auth/logout": {
      "post": {
        "tags": [
          "Authentication"
        ],
        "summary": "User logout",
        "description": "Invalidate current session tokens",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Logout successful",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Logged out successfully"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}